# docs/api/openapi.json
{
  "components": {
    "parameters": {
      "AccountID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "账号 UUID",
        "schema": {
          "type": "string"
        }
      },
      "Bucket": {
        "name": "bucket",
        "in": "query",
        "description": "桶名；账号有默认桶时可省略",
        "schema": {
          "type": "string"
        }
      },
      "ContinuationToken": {
        "name": "continuationToken",
        "in": "query",
        "description": "分页游标",
        "schema": {
          "type": "string"
        }
      },
      "MaxKeys": {
        "name": "maxKeys",
        "in": "query",
        "description": "单页对象数（1-1000）",
        "schema": {
          "type": "integer"
        }
      },
      "Prefix": {
        "name": "prefix",
        "in": "query",
        "description": "前缀（目录路径）",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "资源不存在"
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "example": {
          "error": "account not found"
        },
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
      "ListObjectsResp": {
        "type": "object",
        "example": {
          "objects": [
            {
              "key": "docs/a.txt",
              "size": 17,
              "lastModified": "2026-09-30T05:00:00Z",
              "etag": "\"9c1d2f3a4b5c6d7e\"",
              "storageClass": "STANDARD",
              "isDir": false
            }
          ],
          "commonPrefixes": [
            "docs/"
          ],
          "isTruncated": false,
          "nextToken": ""
        },
        "properties": {
          "commonPrefixes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isTruncated": {
            "type": "boolean"
          },
          "nextToken": {
            "type": "string"
          },
          "objects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectItem"
            }
          }
        },
        "required": [
          "commonPrefixes",
          "isTruncated",
          "nextToken",
          "objects"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "S3C_TOKEN 中的 Bearer token。可通过 S3C_TOKEN_SCOPES 为单个 token 限定 readonly / prefixes（桶或桶内键前缀）/ accounts / expiresAt；越权请求返回 403，过期 token 返回 401。未在 S3C_TOKEN_SCOPES 中登记的 token 为全权。",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "s3client API",
    "version": "v1.0.0",
    "description": "s3client 是面向 S3 兼容对象存储的多账号 Web 控制台。本文档为 /api/* 端点的 OpenAPI 3.0 契约，所有响应均 JSON（除 /api/health 等纯状态端点）。鉴权：Bearer Token（环境变量 S3C_TOKEN，多值逗号分隔）。"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/accounts/{id}/copy-object": {
      "post": {
        "operationId": "copyObject",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs/a.txt",
                "newBucket": "archive",
                "newKey": "a.txt"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "checksumAlgorithm": {
                    "type": "string",
                    "description": "可选；非空时服务端计算并存储全对象校验和（供 verify-checksum 端到端比对）",
                    "enum": [
                      "CRC64NVME",
                      "SHA256",
                      "CRC32C",
                      "SHA1"
                    ]
                  },
                  "ifMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当**目标**对象当前 ETag 匹配时写入"
                  },
                  "ifNoneMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当**目标**对象不存在时写入",
                    "enum": [
                      "*"
                    ]
                  },
                  "key": {
                    "type": "string"
                  },
                  "newBucket": {
                    "type": "string",
                    "format": "可选；省略=同桶"
                  },
                  "newKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key",
                  "newKey"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "copied": "a.txt",
                  "bucket": "archive"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "copied": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "条件字段非法"
          },
          "412": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "条件不满足（PreconditionFailed）"
          }
        },
        "summary": "单文件复制（不删源）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/delete": {
      "post": {
        "operationId": "deleteObjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "keys": [
                  "docs/a.txt",
                  "docs/b.txt"
                ]
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "keys": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "bucket",
                  "keys"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "deleted": 2,
                  "failed": 0
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    },
                    "lastError": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "含 deleted/failed/lastError；S3 逐 key 失败仍返回 200，deleted 只计成功数"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "key 数\u003e1000"
          }
        },
        "summary": "批量删除（≤1000 keys）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/delete-prefix": {
      "post": {
        "operationId": "deletePrefix",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "prefix": "docs/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "prefix"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "deleted": 11,
                  "failed": 0,
                  "truncated": false
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "integer"
                    },
                    "failed": {
                      "type": "integer"
                    },
                    "lastError": {
                      "type": "string"
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "含 deleted/failed/truncated/lastError"
          }
        },
        "summary": "递归删除前缀（同步流式）",
        "tags": [
          "objects"
        ]
      }
    },
    "/api/accounts/{id}/delete-prefix/async": {
      "post": {
        "operationId": "deletePrefixAsync",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "prefix": "docs/"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "prefix": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "prefix"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "example": {
                  "jobId": "6f1e2d3c-4b5a-6789-abcd-ef0123456789",
                  "total": 11,
                  "truncated": false
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobId": {
                      "type": "string"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "truncated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            },
            "description": "jobId"
          }
        },
        "summary": "递归删除前缀（异步）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/head": {
      "get": {
        "operationId": "headObject",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          },
          {
            "$ref": "#/components/parameters/Bucket"
          },
          {
            "description": "",
            "in": "query",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "可选；指定读取某历史版本的元数据",
            "in": "query",
            "name": "versionId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "key": "docs/a.txt",
                  "size": 17,
                  "lastModified": "2026-09-30T05:00:00Z",
                  "etag": "\"9c1d2f3a4b5c6d7e\"",
                  "contentType": "text/plain; charset=utf-8",
                  "storageClass": "STANDARD",
                  "metadata": {
                    "owner": "alice"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "checksums": {
                      "type": "object",
                      "description": "服务端存储的校验和；无校验和或厂商不支持时为 null",
                      "properties": {
                        "crc32c": {
                          "type": "string"
                        },
                        "crc64nvme": {
                          "type": "string"
                        },
                        "sha1": {
                          "type": "string"
                        },
                        "sha256": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string",
                          "description": "FULL_OBJECT | COMPOSITE_*（分段合成，不可全对象比对）"
                        }
                      },
                      "nullable": true
                    },
                    "contentType": {
                      "type": "string"
                    },
                    "etag": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "lastModified": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "metadata": {
                      "type": "object"
                    },
                    "size": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "storageClass": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "summary": "对象元数据",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/mkdir": {
      "post": {
        "operationId": "mkdirObject",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "ifMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当目标对象当前 ETag 匹配时写入"
                  },
                  "ifNoneMatch": {
                    "type": "string",
                    "description": "可选；条件写：仅当目标对象不存在时写入（防并发覆盖）",
                    "enum": [
                      "*"
                    ]
                  },
                  "key": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "created": "docs/",
                  "bucket": "my-bucket"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "bucket": {
                      "type": "string"
                    },
                    "created": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "key/条件字段非法"
          }
        },
        "summary": "新建空文件夹（PUT 空对象）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/objects": {
      "get": {
        "operationId": "listObjects",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          },
          {
            "$ref": "#/components/parameters/Bucket"
          },
          {
            "$ref": "#/components/parameters/Prefix"
          },
          {
            "description": "",
            "in": "query",
            "name": "delimiter",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/MaxKeys"
          },
          {
            "$ref": "#/components/parameters/ContinuationToken"
          },
          {
            "description": "按 key 字典序从该 key 之后开始列举（ListObjectsV2 start-after）；与 continuationToken 互斥使用",
            "in": "query",
            "name": "startAfter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "objects": [
                    {
                      "key": "docs/a.txt",
                      "size": 17,
                      "lastModified": "2026-09-30T05:00:00Z",
                      "etag": "\"9c1d2f3a4b5c6d7e\"",
                      "storageClass": "STANDARD",
                      "isDir": false
                    }
                  ],
                  "commonPrefixes": [
                    "docs/"
                  ],
                  "isTruncated": false,
                  "nextToken": ""
                },
                "schema": {
                  "$ref": "#/components/schemas/ListObjectsResp"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "列对象（含公共前缀 / 分页）",
        "tags": [
          "objects"
        ]
      }
    },
...
    "/api/accounts/{id}/rename": {
      "post": {
        "operationId": "renameObject",
        "parameters": [
          {
            "$ref": "#/components/parameters/AccountID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "bucket": "my-bucket",
                "key": "docs/a.txt",
                "newKey": "archive/a.txt"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "bucket": {
                    "type": "string"
                  },
                  "key": {
                    "type": "string"
                  },
                  "newBucket": {
                    "type": "string",
                    "format": "可选；省略=同桶"
                  },
                  "newKey": {
                    "type": "string"
                  }
                },
                "required": [
                  "bucket",
                  "key",
                  "newKey"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "renamed": "archive/a.txt"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "renamed": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "重命名 / 移动（copy+delete，可跨桶）",
        "tags": [
          "objects"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "/",
      "description": "同源（前端 Vite 代理或后端 SPA fallback）"
    }
  ],
...
}